requirements.txt
asgiref==3.3.4
click==8.0.1
colorama==0.4.4
Django==3.2.4
django-relativedelta==1.1.2
Flask==2.0.1
flask-heroku==0.1.9
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dateutil==2.8.1
pytz==2021.1
radar==0.3
six==1.16.0
sqlparse==0.4.1
Werkzeug==2.0.1
